HSEARCH-3455 Move ElasticsearchClient and related classes to SPI
