--- backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/client/impl/ElasticsearchClientImplementor.java
@@ -1,18 +1,0 @@
-/*
- * Hibernate Search, full-text search for your domain model
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
- */
-package org.hibernate.search.backend.elasticsearch.client.impl;
-
-import org.hibernate.search.backend.elasticsearch.gson.impl.GsonProvider;
-
-/**
- * An interface allowing to configure an {@link ElasticsearchClient}.
- */
-public interface ElasticsearchClientImplementor extends ElasticsearchClient {
-
-	void init(GsonProvider gsonProvider);
-
-}
