+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/client/impl/Paths.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.backend.elasticsearch.client.impl;
 
-import org.hibernate.search.backend.elasticsearch.util.impl.URLEncodedString;
+import org.hibernate.search.backend.elasticsearch.util.spi.URLEncodedString;
 
 /**
  * Useful paths to compose Elasticsearch URLs.
