+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/client/spi/ElasticsearchResponse.java
@@ -4,14 +4,14 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.client.impl;
+package org.hibernate.search.backend.elasticsearch.client.spi;
 
 import com.google.gson.JsonObject;
 
 /**
  * @author Yoann Rodiere
  */
-public class ElasticsearchResponse {
+public final class ElasticsearchResponse {
 
 	private final int statusCode;
