+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/gson/spi/JsonLogHelper.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.gson.impl;
+package org.hibernate.search.backend.elasticsearch.gson.spi;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -19,7 +19,7 @@ import com.google.gson.JsonObject;
  *
  * @author Yoann Rodiere
  */
-public class JsonLogHelper {
+public final class JsonLogHelper {
 
 	private static final JsonLogHelper INSTANCE = create( new GsonBuilder(), true );
