+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/impl/ElasticsearchBackendFactory.java
@@ -15,8 +15,8 @@ import org.hibernate.search.backend.elasticsearch.analysis.model.impl.Elasticsea
 import org.hibernate.search.backend.elasticsearch.cfg.MultiTenancyStrategyConfiguration;
 import org.hibernate.search.backend.elasticsearch.cfg.SearchBackendElasticsearchSettings;
 import org.hibernate.search.backend.elasticsearch.client.impl.ElasticsearchClientFactoryImpl;
-import org.hibernate.search.backend.elasticsearch.client.impl.ElasticsearchClientFactory;
-import org.hibernate.search.backend.elasticsearch.client.impl.ElasticsearchClientImplementor;
+import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchClientFactory;
+import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchClientImplementor;
 import org.hibernate.search.backend.elasticsearch.document.model.impl.esnative.FieldDataType;
 import org.hibernate.search.backend.elasticsearch.document.model.impl.esnative.IndexType;
 import org.hibernate.search.backend.elasticsearch.document.model.impl.esnative.NormsType;
@@ -24,7 +24,7 @@ import org.hibernate.search.backend.elasticsearch.gson.impl.DefaultGsonProvider;
 import org.hibernate.search.backend.elasticsearch.gson.impl.ES5FieldDataTypeJsonAdapter;
 import org.hibernate.search.backend.elasticsearch.gson.impl.ES5IndexTypeJsonAdapter;
 import org.hibernate.search.backend.elasticsearch.gson.impl.ES5NormsTypeJsonAdapter;
-import org.hibernate.search.backend.elasticsearch.gson.impl.GsonProvider;
+import org.hibernate.search.backend.elasticsearch.gson.spi.GsonProvider;
 import org.hibernate.search.backend.elasticsearch.logging.impl.Log;
 import org.hibernate.search.backend.elasticsearch.multitenancy.impl.DiscriminatorMultiTenancyStrategy;
 import org.hibernate.search.backend.elasticsearch.multitenancy.impl.MultiTenancyStrategy;
