+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/orchestration/impl/ElasticsearchStubWorkExecutionContext.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.backend.elasticsearch.orchestration.impl;
 
-import org.hibernate.search.backend.elasticsearch.client.impl.ElasticsearchClient;
+import org.hibernate.search.backend.elasticsearch.client.spi.ElasticsearchClient;
 import org.hibernate.search.backend.elasticsearch.work.impl.ElasticsearchWorkExecutionContext;
 
 /**
